edit filename and del empty files
