+++ lib/main.dart
@@ -4,7 +4,7 @@ import 'package:flutter_riverpod/flutter_riverpod.dart';
 import 'package:ltrc/contants/routes.dart';
 import 'package:ltrc/extensions.dart';
 import 'package:ltrc/views/register_view.dart';
-import 'package:ltrc/views/teachWordView.dart';
+import 'package:ltrc/views/teach_word_view.dart';
 // import 'package:ltrc/views/units_view.dart';
 import 'package:ltrc/views/words_view.dart';
